requirements.txt
# Ecosystem Mapper Dependencies

# API clients
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.12.0  # For OpenRouter compatibility

# GitHub API
PyGithub>=2.1.1

# Tavily search
tavily-python>=0.3.0

# Data processing
pydantic>=2.5.0

# CLI
click>=8.1.7

# Date handling
python-dateutil>=2.8.2
